Create base files for mobile view, currently WIP
