+++ src/components/ExportCSV/ExportCSVForm/ExportCSVForm.module.css
@@ -105,3 +105,113 @@
   color: red;
   font-size: 13px;
 }
+
+@media screen and (max-width: 768px) {
+  .csv-form-wrapper {
+    flex-basis: 42%;
+    width: 100%;
+    margin: 1.4em 0em 1.4em;
+  }
+
+  .export-csv-form {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+
+  .section-wrapper {
+    width: 400px !important;
+  }
+
+  .filter-label-select {
+    width: 400px !important;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .input-drop-down {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 200px;
+    gap: 5px;
+  }
+
+  .organically-radio-group {
+    width: 200px;
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .date-picker {
+    font-size: var(--chakra-fontSizes-md);
+    height: 2.5rem;
+    width: 200px;
+    border: 1px solid rgb(226, 232, 240);
+    border-radius: 0.375rem;
+    padding-left: 1rem;
+  }
+
+  .date-picker-error {
+    font-size: var(--chakra-fontSizes-md);
+    height: 2.5rem;
+    width: 200px;
+    border: 1px solid #e53e3e;
+    box-shadow: 0 0 0 1px #e53e3e;
+    border-radius: 0.375rem;
+    padding-left: 1rem;
+  }
+
+  .date-range {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+
+  .date-range-label {
+    font-size: 14px;
+    color: #9c9c9c;
+  }
+
+  .filter-section-wrapper {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .filter-option-wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+
+  .filter-choices {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .box-detail-checkboxes {
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    height: 100px;
+  }
+
+  .box-detail-header {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .csv-form-labels {
+    font-size: 20px;
+    font-weight: 500;
+  }
+
+  .error-message {
+    color: red;
+    font-size: 13px;
+  }
+}
